requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-appconf==1.0.6
django-imagekit==5.0.0
django-ipware==7.0.1
djangorestframework==3.14.0
idna==3.6
lxml==5.2.1
pilkit==3.0
pillow==10.2.0
python-ipware==3.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1